requirements.txt
behave~=1.2.6
python-dotenv~=1.1.0
requests~=2.32.3
allure-python-commons~=2.14.2
urllib3~=2.4.0
allure-behave~=2.14.2